moved md-flex sources to md-flexible/src
