+++ examples/md-flexible/src/Generator.h
