+++ examples/md-flexible/src/Objects/CubeGauss.h
