+++ examples/md-flexible/tests/GeneratorsTest.cpp
@@ -6,12 +6,11 @@
 
 #include "GeneratorsTest.h"
 
-#include "Generator.h"
-#include "PrintableMolecule.h"
 #include "autopas/AutoPas.h"
 #include "autopasTools/generators/GridGenerator.h"
 #include "autopasTools/generators/RandomGenerator.h"
-#include "parsing/YamlParser.h"
+#include "src/parsing/YamlParser.h"
+#include "src/Generator.h"
 #include "testingHelpers/commonTypedefs.h"
 
 // the following test only work if testParsing.yaml is well set
