+++ examples/md-flexible/tests/ThermostatTest.cpp
@@ -6,8 +6,8 @@
 
 #include "ThermostatTest.h"
 
-#include "Thermostat.h"
 #include "autopasTools/generators/GridGenerator.h"
+#include "src/Thermostat.h"
 
 void ThermostatTest::initContainer(AutoPasType &autopas, const Molecule &dummy, std::array<size_t, 3> particlesPerDim) {
   constexpr double particleSpacing = 1.;
